# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vancouver Competitive LoL Project API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AdminLoginRequest {
  email: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
  email?: string;
}

export interface AdminStats {
  interests: number;
  events: number;
  registrations: number;
  matches: number;
  vods: number;
  players: number;
  seasons: number;
}

export interface InterestSubmission {
  id: number;
  riotId: string;
  discordUsername: string;
  currentRank: string;
  city: string;
  preferredFormat: string;
  availability: string;
  hasTeam: boolean;
  willingWithoutPrize: boolean;
  notes?: string | null;
  createdAt: string;
}

export interface CreateInterestRequest {
  riotId: string;
  discordUsername: string;
  currentRank: string;
  city: string;
  preferredFormat: string;
  availability: string;
  hasTeam: boolean;
  willingWithoutPrize: boolean;
  notes?: string | null;
}

export interface Event {
  id: number;
  title: string;
  slug: string;
  format: string;
  eventDate: string;
  registrationStatus: string;
  shortDescription: string;
  fullDescription?: string | null;
  rulesSummary?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Match {
  id: number;
  eventId?: number | null;
  eventTitle?: string | null;
  matchTitle: string;
  sideAName: string;
  sideBName: string;
  winnerName: string;
  score?: string | null;
  format?: string | null;
  vodUrl?: string | null;
  playerAId?: number | null;
  playerBId?: number | null;
  playerAEloBefore?: number | null;
  playerAEloAfter?: number | null;
  playerBEloBefore?: number | null;
  playerBEloAfter?: number | null;
  seasonId?: number | null;
  isPlayoff: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface VodEntry {
  id: number;
  eventId?: number | null;
  eventTitle?: string | null;
  title: string;
  format?: string | null;
  playerNames?: string | null;
  roleTag?: string | null;
  notes?: string | null;
  videoUrl: string;
  playerId?: number | null;
  playerRiotId?: string | null;
  champion?: string | null;
  opponentChampion?: string | null;
  position?: string | null;
  patch?: string | null;
  playerEloAtTime?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface EventDetail {
  id: number;
  title: string;
  slug: string;
  format: string;
  eventDate: string;
  registrationStatus: string;
  shortDescription: string;
  fullDescription?: string | null;
  rulesSummary?: string | null;
  createdAt: string;
  updatedAt: string;
  matches: Match[];
  vods: VodEntry[];
}

export interface CreateEventRequest {
  title: string;
  slug: string;
  format: string;
  eventDate: string;
  registrationStatus: string;
  shortDescription: string;
  fullDescription?: string | null;
  rulesSummary?: string | null;
}

export interface EventRegistration {
  id: number;
  eventId: number;
  eventTitle?: string | null;
  riotId: string;
  discordUsername: string;
  currentRank: string;
  city: string;
  availabilityConfirmation: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateRegistrationRequest {
  eventId: number;
  riotId: string;
  discordUsername: string;
  currentRank: string;
  city: string;
  availabilityConfirmation: string;
  notes?: string | null;
}

export interface CreateMatchRequest {
  eventId?: number | null;
  matchTitle: string;
  sideAName: string;
  sideBName: string;
  winnerName: string;
  score?: string | null;
  format?: string | null;
  vodUrl?: string | null;
  playerAId?: number | null;
  playerBId?: number | null;
  seasonId?: number | null;
  isPlayoff?: boolean | null;
}

export interface CreateVodRequest {
  eventId?: number | null;
  title: string;
  format?: string | null;
  playerNames?: string | null;
  roleTag?: string | null;
  notes?: string | null;
  videoUrl: string;
  playerId?: number | null;
  champion?: string | null;
  opponentChampion?: string | null;
  position?: string | null;
  patch?: string | null;
  playerEloAtTime?: number | null;
}

export interface Player {
  id: number;
  riotId: string;
  discordUsername: string;
  currentElo: number;
  peakElo: number;
  wins: number;
  losses: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlayerRequest {
  riotId: string;
  discordUsername: string;
  currentElo?: number | null;
  isActive?: boolean | null;
}

export interface PlayerProfile {
  id: number;
  riotId: string;
  discordUsername: string;
  currentElo: number;
  peakElo: number;
  wins: number;
  losses: number;
  isActive: boolean;
  recentMatches: Match[];
  vods: VodEntry[];
  createdAt: string;
  updatedAt: string;
}

export interface Season {
  id: number;
  name: string;
  status: string;
  startDate: string;
  endDate: string;
  eloResetFactor: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSeasonRequest {
  name: string;
  status?: string | null;
  startDate: string;
  endDate: string;
  eloResetFactor?: string | null;
}

export interface LadderEntry {
  rank: number;
  id: number;
  riotId: string;
  discordUsername: string;
  currentElo: number;
  peakElo: number;
  wins: number;
  losses: number;
  winRate: number;
}

export interface LadderResponse {
  season?: Season | null;
  entries: LadderEntry[];
}

export interface VodTimestamp {
  id: number;
  vodId: number;
  label: string;
  seconds: number;
  type: string;
  createdAt: string;
}

export interface CreateVodTimestampRequest {
  label: string;
  seconds: number;
  type?: string | null;
}

export interface VodDetail {
  id: number;
  eventId?: number | null;
  eventTitle?: string | null;
  title: string;
  format?: string | null;
  playerNames?: string | null;
  roleTag?: string | null;
  notes?: string | null;
  videoUrl: string;
  playerId?: number | null;
  playerRiotId?: string | null;
  champion?: string | null;
  opponentChampion?: string | null;
  position?: string | null;
  patch?: string | null;
  playerEloAtTime?: number | null;
  timestamps: VodTimestamp[];
  relatedVods: VodEntry[];
  createdAt: string;
  updatedAt: string;
}

export type ListRegistrationsParams = {
  eventId?: number;
};

export type ListMatchesParams = {
  eventId?: number;
  format?: string;
  search?: string;
};

export type ListVodsParams = {
  eventId?: number;
  format?: string;
  roleTag?: string;
  search?: string;
  champion?: string;
  opponentChampion?: string;
  position?: string;
  patch?: string;
  eloMin?: number;
  eloMax?: number;
};