# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vancouver Competitive LoL Project API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /interests:
    get:
      operationId: listInterests
      tags: [interests]
      summary: List all interest submissions (admin)
      responses:
        "200":
          description: List of interest submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterestSubmission"
    post:
      operationId: createInterest
      tags: [interests]
      summary: Submit general interest form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterestRequest"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterestSubmission"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /interests/{id}:
    delete:
      operationId: deleteInterest
      tags: [interests]
      summary: Delete an interest submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{slug}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/edit:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/delete:
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    AdminLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    AdminMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        email:
          type: string
      required:
        - authenticated
...
    InterestSubmission:
      type: object
      properties:
        id:
          type: integer
        riotId:
          type: string
        discordUsername:
          type: string
        currentRank:
          type: string
        city:
          type: string
        preferredFormat:
          type: string
        availability:
          type: string
        hasTeam:
          type: boolean
        willingWithoutPrize:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - riotId
        - discordUsername
        - currentRank
        - city
        - preferredFormat
        - availability
        - hasTeam
        - willingWithoutPrize
        - createdAt
    CreateInterestRequest:
      type: object
      properties:
        riotId:
          type: string
        discordUsername:
          type: string
        currentRank:
          type: string
        city:
          type: string
        preferredFormat:
          type: string
        availability:
          type: string
        hasTeam:
          type: boolean
        willingWithoutPrize:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - riotId
        - discordUsername
        - currentRank
        - city
        - preferredFormat
        - availability
        - hasTeam
        - willingWithoutPrize
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        format:
          type: string
        eventDate:
          type: string
        registrationStatus:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
          nullable: true
        rulesSummary:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - format
        - eventDate
        - registrationStatus
        - shortDescription
        - createdAt
        - updatedAt
    EventDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        format:
          type: string
        eventDate:
          type: string
        registrationStatus:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
          nullable: true
        rulesSummary:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        vods:
          type: array
          items:
            $ref: "#/components/schemas/VodEntry"
      required:
        - id
        - title
        - slug
        - format
        - eventDate
        - registrationStatus
        - shortDescription
        - createdAt
        - updatedAt
        - matches
        - vods
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        format:
          type: string
        eventDate:
          type: string
        registrationStatus:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
          nullable: true
        rulesSummary:
          type: string
          nullable: true
      required:
        - title
        - slug
        - format
        - eventDate
        - registrationStatus
        - shortDescription
...
    Match:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
          nullable: true
        eventTitle:
          type: string
          nullable: true
        matchTitle:
          type: string
        sideAName:
          type: string
        sideBName:
          type: string
        winnerName:
          type: string
        score:
          type: string
          nullable: true
        format:
          type: string
          nullable: true
        vodUrl:
          type: string
          nullable: true
        playerAId:
          type: integer
          nullable: true
        playerBId:
          type: integer
          nullable: true
        playerAEloBefore:
          type: integer
          nullable: true
        playerAEloAfter:
          type: integer
          nullable: true
        playerBEloBefore:
          type: integer
          nullable: true
        playerBEloAfter:
          type: integer
          nullable: true
        seasonId:
          type: integer
          nullable: true
        isPlayoff:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - matchTitle
        - sideAName
        - sideBName
        - winnerName
        - isPlayoff
        - createdAt
        - updatedAt
...
    VodEntry:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
          nullable: true
        eventTitle:
          type: string
          nullable: true
        title:
          type: string
        format:
          type: string
          nullable: true
        playerNames:
          type: string
          nullable: true
        roleTag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        videoUrl:
          type: string
        playerId:
          type: integer
          nullable: true
        playerRiotId:
          type: string
          nullable: true
        champion:
          type: string
          nullable: true
        opponentChampion:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        patch:
          type: string
          nullable: true
        playerEloAtTime:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - videoUrl
        - createdAt
        - updatedAt
...